APIs with authentication #21

+++ todo-list/backend/controllers/noteController.js
@@ -3,16 +3,16 @@ const Note = require('../models/noteModel');
 
 //@desc Get all notes
 //@route GET /api/notes
-//@access public
+//@access private
 
 const getNotes = asyncHandler(async (req, res) => {
-    const notes = await Note.find({});
+    const notes = await Note.find({user_id: req.user.id});
     res.status(200).json(notes);
 });
 
 //@desc create new note
 //@route POST /api/notes
-//@access public
+//@access private
 
 const createNote = asyncHandler(async (req, res) => {
     const {text, completed} = req.body;
@@ -22,20 +22,26 @@ const createNote = asyncHandler(async (req, res) => {
     }
     const note = await Note.create({
         text, 
-        completed
+        completed,
+        user_id: req.user.id,
     }); 
     res.status(201).json(note);
 });
 
 //@desc Update note
 //@route PUT /api/notes/:id
-//@access public
+//@access private
 
 const updateNote = asyncHandler(async (req, res) => {
     const note = await Note.findById(req.params.id);
     if (!note) {
         res.status(404);
         throw new Error("Note not found");
+    }
+
+    if(note.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("You are not authorized to update this note");
     }
 
     const updatedNote = await Note.findByIdAndUpdate(
@@ -48,7 +54,7 @@ const updateNote = asyncHandler(async (req, res) => {
 
 //@desc delete note
 //@route DELETE /api/notes/:id
-//@access public
+//@access private
 
 const deleteNote = asyncHandler(async (req, res) => {
     const note = await Note.findById(req.params.id);
@@ -56,6 +62,12 @@ const deleteNote = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error("Note not found");
     }
+
+    if(note.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("You are not authorized to delete this note");
+    }
+
     await note.deleteOne();
 
     res.status(200).json(note);
